Make updatingCacheWriteWorkingDirectory faster by checking out files individually
I don't understand exactly what's happening, but: when calling checkoutFilesToWorkingDirectory (and therefore git_checkout_index) with multiple paths, something happens.
It seems that libgit2 finds the first common ancestor of those files; and then iterates over the complete tree of this ancestor's children, including ignored ones!

So, for instance, if the repo contains a folder with a huge number of files, cache/node_modules/:
- Passing index.js as path will be fast: there's just one file, so no iteration
- Passing index.js and views/welcome.js as paths will be slow: their ancestor is /, which contains cache/node_modules/
- Passing views/welcome.js and views/discover.js will be fast: their ancestor is views/
- Passing cache/ducks.json and cache/embeds/toots.json will be slow
- Passing cache/embeds/toots.json and cache/embeds/friends.json will be fast
- Passing cache/embeds/toots.json, cache/embeds/friends.json, and views/welcome.js will be slow (ancestor is /)

I don't understand why libgit2 would explore the tree (when it's supposed to only consider the specified paths), and ignore the fact that node_modules is ignored. I *think* passing GIT_CHECKOUT_DISABLE_PATHSPEC_MATCH solved the performance issue, but also introduced strange differences in behavior.

So, a very simple fix for this issue is to call checkoutFilesToWorkingDirectory repeatedly, once with each path, instead of calling it with all the paths at once.