updateWorkingDirectoryCache: Revert workdir cache index reuse
This hurts performance.

Turns out that plain doesn't work: some parts of Retcon *store* the workdir content index, after reading it, and expect it not to change. (e.g. the undo point system)

Changing that assertion in Retcon would be a bunch of error-prone work, and might not result in any performance improvement in the end.

(thankfully, this is caught by Retcon's tests)