Workdir cache: Don't write repository index
This finally removes the need to write the repo index, to update our workdir cache! This should completely do away with the interference that would happen between Retcon and other tools. (e.g. confusing the Xcode gutter Git status indicators, or things getting stage out of the blue)

This relies on git_repository_set_index, which allows us to (here, temporarily) designate a different index as the repository's index. This in turns allows us to use status lists and staging methods on our cache index, avoiding the need to write the actual repository index file.

This change also:
- Simplifies the code: no more juggling the rpeo and cache indexes
- Probably makes the code faster, as we don't need to copy between these two indexes anymore